More Changes
Note to Scott, by sending the person over part and parcel, and observing
all, set to Firebase would not work un-entered fields were undefined. I
believe the problem has to do with how the keys are accessed.  We dont
want to actually write to the same key right there, I don't think.  I
could change it to an update.