List items in a neon animation infinite list filled from database with "calls"
Phase one of a smooth integration of a single element that can be
dropped anywhere which holds its own meta, has input search attached as
well as independent search bar.

Animation still uses preset images.  We have the option with animation,
however, of changing the size and scope of the view change.

1. We can have a "page" or set of pages for each call in that is
restricted to the single box element.
2. We can create a view shift to full screen "page" view with addition
of button.
3. Animation does not make a permanent rendering, meaning that we should
save on loads and lags.
4. Plan to add meta for call entry, and search, so we have an effective
single local buffer which will speed search.
5. May use local storage to hold background, database checking, call
list version to eliminate search and send data lags we are currently
experiencing with smaller call set, and low data.

A month of calls should be 2 megs; WITH DATA... 10MEGS? 7G COMPRESSION
MAY HELP, LISTENED ABOUT IT LAST NIGHT.

that's where I am.

Enjoy,