These are my Changes to Inputs
Have been attempting to create an integrated call log and input system.
The inputs fc-paperInput is in the folder with the listview work.

The attempt was to create a set of predefined "names" for the inputs,
that would then return placeholders and labels which would make the
inputs fit perfectly into a call-script and related message text.  As
you can see I have had some issues, not the least of which includes how
to return multiple values in an initialization function.

As of right now the inputs work to add a call, but the input message
text above the call list, as well as the inputted call list message are
sort of a mess.

I am registering this commit only to log it for your assistance.