These moved the Inputs to the top
Inputs are working, and creating a narrative call message.  I am trying
to get the inputs to be name-able at creation, but defining them with
the name used in a markup of a parent element has proven difficult.  I
am unsure how to use the "var element= new FcPaperInput(name)" to define
the elements in the markup.  help would be awesome