Here are Scotts upgrades.
implemented basic student view.  Not constructor.  Implemented
constructor and Factory Implementation for the clock.  He had it working
on individual hard coded instantiations of id for the clock.  Later we
did a hard coded and a factory impl and had both work.  Will discuss
advantages from an input and self-generating code pov.  John--to house
morning?