This is the latest on a redesign.
I havent done all the animation yet, but I have it ready.  It shouldnt
take that long.  Once I get it on one card I will have the remainder.
We already did the animation for the exposure on the other cards we had,
so I can bring that back in easy.  The small details are the real
problem.  All the awesome little css tricks that google people put in.
Amazing really. It was nice to just be able to read it.