Please See the Selectelement, the course detail card, and the fc-assistant
All are moving towards completion.  They each need small tweeks.

The card needs to be wired to db;

the assistant needs a slight redesign in parts: overlay for
paper-dialog; a display none or hidden property for the open window Js
Call.  (which doesnt work). And then I wanna make it drag and drop.

The selector, well, we know the issue with that one scott... well we
know we dont know the issue.;