Work on selects, some commits and merge reversions
This commit holds my work on the inputs, selectors, assistant, redesign
of interface, and attempted (but failed) merge of Scotts global selector
constructor into the intake system.

This commit stands at a stumbling block for me.  I need to have selects
for classes, teachers, times and tests--based on the age/grade level of
students.

These selects are chained and must be entered via call center level
admin staff, meaning we need an interface that regularizes inputs for
the creation of courses, new employee teachers, locations, and course
offerings.

As was evident from last go round, an open input system is a no-no.
Similarly, a strictly hard-coded predefined system lacks the needed
versatility for two reasons:

(1) In House: In house over the next 21 months--if we do it right--we
will see an explosion of opportunity.  Between now and 2020, the psat
8.9, psat 10, psat msnqt and 1600 sat will become placement metrics for
students, dividing line salary and retention determiners for teachers
and ranking determiners for schools.  As a result, our opportunity to
offer deep learning, concentrated work for extended periods (years) and
develop student talent from sixth or seventh grade through high school
will be exponentially greater.

This is a dream come true for me and focused.  We have always wanted to
become long term coaches, mentors, advisors and friends of the next
generation of leaders and thinkers.

We have that opportunity now.  We do not, however, know the limits,
rubrics or specs of the coming lower level tests, and we are unfamiliar
with the particular needs of the future clients to adjust to these
unknown future metrics.

We can adjust quickly material-wise, but we should also be able to
adjust quickly system-wise.  A versatile system from day one will make
our lives at day 400, when material creation is at an extreme premium,
that much easier and more efficient.  Let's not short-term this.
Let's--for once--plan and act to win the war not the race to the first
battle.

2.  The versatility mentioned above makes the system we are building
eminently sellable across platforms.  I am not specifically thinking of
the "second/third business" here, it just seems a natural compliment,
meaning: do it right to not fuck us, do it well enough we love it and
gain the 30-50% bump efficiency could represent, and then we also happen
to gain revenue stream two which is 5x revenue stream 1 if exploited and
far less in need of service...

just say, let's be smart and ahead of the curve once... and if we have
to choose that once, choose when the tasks are the MOST FISCALLY, TIME
PLANNING and EASE OF ALL FUTURE ENDEAVORS beneficial.

It's the only reasonable option.  ;-)