Check out the Experiments Page
The Selector is two/three elements.  A selector as a factory
Implimentation a holder that calls the selector and defines the scope,
and then the host, which provides an id in this instance but that is not
needed.

The final implementation, to create any selector, or really any menu
requires only a path to a database, and then some IDs and
labels/placeholders.  I will add filters and linking/chaining between
elements next.  This is pretty cool in my book and what I have been
hoping to do for two months.