Mid Stretch Work
Have questions about the animation.

Previously, I had a no bug system, using six "chip cards", two regular
cards and then another regular card that animated in a different local.

The errors that arose in the log about two weeks ago popped to a problem
when I went to 12 cards, and had a pair of "merge issues", of the type
from which we have all suffered.

I check in, check out system--would likely work best... Although, we
truly have not intersected much file wise lately.

The best way to proceed is unknown, however, I would love a hand on this
animation.  If we can get it working again, using triads of sets of
elements, the looping can be pretty cool.

Thanks

Jason