Hey Guys... Next on the list is wiring up the inputs.  Settling on an animation or a hide and show for the data assistant.  Settling on a nice animation for the left right slide of the sign up form.
Then we hit the real work.  Take the second search and drag from scotts work.

Integrate the actual payment from the march system, and continue to plow through the class building, etc.

I need a time API for the cards.

We still need to get calendaring, and email done.  I would like to get off the input page before the weekend.

To do that we just have to settle in on a few decisions.  I think we have an awesome chance.

Thanks.

Jason