This is my still problematic attempt at a drag and drop flexible grid which will carry its data to the new drag and drop locations.
My issues related to the "griditemArray[counter].name" assignment.

At ready I had assign griditemArray[counter] = {"name"=counter}; this created an array of objects.

Using the this.push(assignment) created a readable array.

This solved the problem I was haing in the numbers.  Now I have assigned two values to the cards.

the label value is the index value in the array... the input value is the name value in the array.

It now works such that the label showing is the index of the item array,

and the input value is the value of the name subproperty of the array at the index number

I have named them the same number, so that when we get the drag and drop and then the
array mutation working, the label numbers should change, but the input values should not.

I would love as much help as possible on this implimentation, as I believe that this implimentation,
combined with a proper fanned-out data structure is
1. the point of using polymer
2. will give us a powerful design and data development tool
3. will vastly accellerate production
4. will make the data sets easier to manage in terms of presentation and change
and
5. will be cool and will translate easily into a cool vocab word block game that will rapidly increase site traffic.