Okay... Reasonale progress on the generic cards and inputs and saving back. Had a minor step backwards, but am recovered.
Am trying to test the ability to both drag and drop accross grids

Am curious as to its effect on routing dat.  If we have a
"key system, with fan-our or without" we still could just send the
"this data yes key to a location and that is the read.

If keys work at all, this sort of linking shouldnt be that tough, with some thought fanning out
miight be a nice intermeddiate solution.

Highlights on this push are the cards in test 1 and 2.  The next push promises far more.

Thanks...
Jason