I would only take the data administrator from this push; It's the only thing that is running well.  It has the word arrays.  It has the passage in it.  I want to get the dictionary stuff up and running then just sort of kill the array storage and measuring.
I believe we can nail this thing if we just take apart each piece one segment at a time.

The CSS answer might be to just use flex, and "know where the items will be"
I can make it dynamic with font sizes if you all want. Its just a property change on the canvas; We could analyze font too if you want.