Did not get quite as far as I wanted to.  I have a dom-repeat of spans that have marked in ids, the paragraph breaks.
I attempted to flick a hidden property on dom repeat items to get the paragraphs marked this was a fail.

I believe a subsequent attempt after the fact, however, will work--like right after rendering... in a parent ready function.
Similarly, have a method for activating some content editable on cells and making them essentially into inputs.  Will need some help extending these elements.

if we do that, with the binding (which I showed worked with paper inputs--

well then we have the word array, bound, editable.

add the drag and drop behavior and the new APIs, plus--lastly--the sizing for the splitter and the line numbers and we finished goal 1. Yeah!