Mostly working three element line numbers and passages, single and double columns.
Pluses:
1.speed: live dynamic-you see them dance--updates on words and line numbers.
2. re-usability: it forced a load set from on-child-added in the holder/viewer parent element.  this pattern--extended, points us towards classes, behaviors, and more generality, shared code!
3. flexibility: the pieces click in and out.  Bonus. one breaks all do not break.

negatives:
1. bug: have a bug on line numbers, was not thorough enough on which anchor is where, so the line numbers can blink out when a id vanishes and update not fast enough.  behavior or class or measuring add on behavior fixes.

any debugging from y'all would likely do the same.

2. save well... or icky... although i can fix css wise

this is not a modular fail, its just that data in has to be the same always if data out is to look top notch.

TOTAL: we are moments away from a WOW set here.

we need: a PROMISE to make sure loads are cool.

I cheated around it for eth single, and could do the same for the double... shitty chaotic coding; promise ++ classing will help

its time to move to the drawer!!! yeah!!! build me a test, folks