made big push:
have working dragable and resizable "cards"
made with a paper header panel.  I placed the formatted test header
into the header content are.  It looks pretty good, but the point of the move was to
have a header panel that would not move or scroll

will need to take a look at all the scrolling.

Has been thrown in willy nilly

in general through, great progress.