have a hide and reveal switch that works on the twin displayer, utilizing the splitter.
had a version which did toggling right and left but it was
hack and slash, erratic.  So i removed it.

Need help on animation in a repeater.

Goal is to get his page set up to repeat 4-8 times to build an area.

and to use the animation for
transitions rather than traditional tabs or
shitty hide reveal.