First run to database and back of a Class-Layout saver.
With the right content pieces,
it will allow us to build almost
everything with just class and type switching;

models up on the site. You can play with them.

To add content, also piecing together
an in-out passage singleton that uses es-6;

it already takes passages in; it does the word array through
firepad; you can see a bad load in a passage displayer
prior to saving.

I need two or three more pieces to have that up and running;

when it runs;
I will combine it with the layout maker for questions;

big commit;  long day and night;