refactored the passage saving panel for passage saving panel page two:
changed the 1100 lines of code into 310 by turning
the first half of the saving and the processing of the passages into a behavior.

the passage saving panel is still able to load all of its pages and will shortly load from firebase.
page two uses an animated list.  My goal here is to have the list
represent questions; the blow up will be the question in context and with others
of its kind.

the center area has a firepad and and drop down and text areas.

the text areas have a feature i would love to "reverse"

input data we throw in automatically gets tossed into one of two bins:

wild card (replace it... perhaps useful as a place holder between units o
useful info) or lock piece....

one of the primary problems I am having is representing that sort of inform
ation on the screeen or in the database.

it is my belief that multiple fan out views will clarify.

as will multiple, array driven class layouts.