Placed es-6 page builder modules into animated declarative pages on home page.
will next move the question template from page one two page two.
change the inputs from the textarea to individual bound inputs in the entry areas,
all based on question patterns i see.

I will then build a template for each of the patterns, and connect said template
to a drop down.

Tabulation of pattern types will be included, such that balancing test creation will be easier.

spread of questions should be included as well.

this will then be logged to the data base with the appropriate related test.

i then will attach highlighting in four colors to mark test for irac and crec.

this will represent the completion of stage 1